move olm utils to their own file
